+++ libs/vedana-etl/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2025-09-04
+
+* Covered get_grist_data / filter_grist_data with tests
+* More fixes in processing "reference" type columns for both links and attrs
+
 # 2025-09-02
 
 * Refactor to var-style Table declaration in pipeline