+++ projects/vedana/CHANGELOG.md
@@ -1,6 +1,9 @@
 # WIP
 
 * (vedana-etl) Refactor to var-style Table declaration in pipeline
+* (vedana-etl) covered get_grist_data / filter_grist_data with tests
+* (vedana-etl) more fixes in processing "reference" type columns for both links
+  and attrs
 
 # 2025.08.29-dev.19